requirements.txt
python-dotenv==0.6.1
simplejson==3.10.0
websocket-client==0.40.0
Rx==1.5.7
slacker==0.9.30
python-jenkins==0.4.13

